refactor: callbacks refactoring
